Fix section symbol in regex when compiled with Angular
For some reason, Angular translates the section symbol in the property name regex to `\xa7`, I assume to escape it, however this causes the regex to match against any `a` in the property name. In this commit I fixed it so that the regex already includes the escaping, so Angular doesn't try to escape it. We are using this in our system and it works great.